Eliminated var, discarded unused params, extracted setting of title and description to middleware, added checking for route.body params for safety

+++ routes/admin.js
@@ -8,17 +8,17 @@ const path = require("path");
 const imageProcessor = require("../image-processing");
 const bcrypt = require("bcrypt");
 const globals = require("../globals.js");
-var unirest = require("unirest");
+let unirest = require("unirest");
 const config = require("config");
 const logger = require("../logger");
-var xl = require('excel4node');
+let xl = require('excel4node');
 
 function translate(text, source, dest) {
   //500000 requests a month !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
   return new Promise((resolve, reject) => {
     if (source == "be") source = "ru";
     if (dest == "be") dest = "ru";
-    var req = unirest(
+    let req = unirest(
       "POST",
       "https://microsoft-translator-text.p.rapidapi.com/translate"
     );
@@ -69,7 +69,6 @@ async function DeleteImageById(nameId, folder) {
   return fs.unlink(imgToDel);
 }
 
-
 async function MakeDefaultImage(newId, folder) {
   let src = path.join(__dirname, "..", folder, "placeholder.jpg");
   let dest = path.join(__dirname, "..", folder, newId + ".jpg");
@@ -81,7 +80,7 @@ async function SaveTmpPoster(tmpfile, dstFolder, newId, thumbnailFolder) {
   let dir = path.dirname(tmpfile);
   let src = path.join(dir, name + ".jpg");
   let dst = path.join(__dirname, "../", dstFolder, newId + ".jpg");
-  var dstThumbnail;
+  let dstThumbnail;
   if (thumbnailFolder) {
     dstThumbnail = path.join(__dirname, "../", thumbnailFolder, name + ".jpg");
   }
@@ -116,7 +115,7 @@ async function PosterUpload(fileToUpload, folder, id, imageProcessorFunction) {
 }
 
 function FilesToArray(files) {
-  var filesArray = [];
+  let filesArray = [];
   if (!Array.isArray(files.files)) {
     filesArray.push(files.files);
   } else {
@@ -135,10 +134,9 @@ router.use(function (req, res, next) {
 });
 
 //Routes
-router.get("/", async function (req, res, next) {
+router.get("/", async function (_req, res, next) {
   try {
     let [results] = await db.query("CALL STAT()");
-    var title = "Admin" + " | " + res.__("title");
     let stat = results[0][0];
     let galleryFiles = await fs.readdir("./static/img/gallery");
     let disksFiles = await fs.readdir("./static/img/disks");
@@ -146,15 +144,14 @@ router.get("/", async function (req, res, next) {
     stat.gallery_count = galleryFiles.length;
     stat.disks_count = disksFiles.length;
     stat.press_count = pressFiles.length;
-    res.render("admin/admin", { title, stat });
+    res.render("admin/admin", { stat });
   } catch (error) {
     next(error);
   }
 });
 
-router.get("/login", (req, res) => {
-  var title = "Login" + " | " + res.__("title");
-  res.render("admin/login", { title });
+router.get("/login", (_req, res) => {
+  res.render("admin/login");
 });
 
 router.post("/login", urlencodedParser, (req, res) => {
@@ -176,7 +173,6 @@ router.get("/logout", function (req, res) {
 
 router.get("/concerts", async (req, res, next) => {
   try {
-
     let itemCount = config.get("paginationSize").admin;
     let currentPage = Number(req.query.page)||1;
     let offset=(currentPage-1)*itemCount;
@@ -207,14 +203,11 @@ router.get("/concerts", async (req, res, next) => {
   }
 });
 
-
-
-
 router.post("/concerts/delete", urlencodedParser, async (req, res, next) => {
   try {
     let id=Number.parseInt(req.body.id);
     await db.query(`DELETE FROM concerts WHERE id=${id}`);
-    DeleteImageById(req.body.id, "/static/img/posters/").then(() => {
+    DeleteImageById(id, "/static/img/posters/").then(() => {
       res.redirect("/admin/");
     });
   } catch (error) {
@@ -273,9 +266,8 @@ router.get("/concerts/export", urlencodedParser, async (req, res, next) => {
   }
 });
 
-router.post("/concerts/add", urlencodedParser, async (req, res, next) => {
+router.post("/concerts/add", urlencodedParser, async (_req, res, next) => {
   try {
-
     let [results] = await db.query(
       `INSERT INTO concerts VALUES (0,'',DATE_FORMAT(NOW() + INTERVAL 1 DAY, '%Y-%m-%d %H:%i:00'),'', '','',1)`
     );
@@ -295,7 +287,7 @@ router.post("/concerts/copy", urlencodedParser, async (req, res, next) => {
       `INSERT INTO concerts VALUES (0,'${concertToCopy.title}',DATE_FORMAT(NOW() + INTERVAL 1 DAY, '%Y-%m-%d %H:%i:00'),\
       '${concertToCopy.description}', '${concertToCopy.place}','${concertToCopy.ticket}',1)`
     );
-    await fs.copyFile(`static/img/posters/${req.body.id}.jpg`,`static/img/posters/${results.insertId}.jpg`);
+    await fs.copyFile(`static/img/posters/${id}.jpg`,`static/img/posters/${results.insertId}.jpg`);
     res.redirect("/admin/");
   } catch (error) {
     next(error);
@@ -303,13 +295,11 @@ router.post("/concerts/copy", urlencodedParser, async (req, res, next) => {
 });
 
 router.post("/concerts/edit", urlencodedParser, async (req, res, next) => {
-
-  var description = viewhelpers.EscapeQuotes(req.body.description);
   let hidden = typeof req.body.hidden == "undefined" ? 0 : 1;
-
+  let description = viewhelpers.EscapeQuotes(req.body.description);
   try {
     if (!viewhelpers.isDate(req.body.date)) throw new Error("Invalid date");
-    var date = req.body.date.slice(0, 19).replace("T", " ");
+    let date = req.body.date.slice(0, 19).replace("T", " ");
     let id=Number.parseInt(req.body.id);
     await db.query(
       `UPDATE concerts SET title = '${req.body.title}', \
@@ -360,7 +350,6 @@ router.get("/news", async (req, res, next) => {
     let maxCount=countDbResult[0].count;
     let pages=viewhelpers.usePagination("/admin/news",currentPage,maxCount,itemCount);
 
-
     events.forEach((element) => {
       element.text = viewhelpers.UnescapeQuotes(element.text);
       element.date = viewhelpers.DateToISOLocal(element.date);
@@ -370,8 +359,6 @@ router.get("/news", async (req, res, next) => {
     next(error);
   }
 });
-
-
 
 router.post("/news/delete", urlencodedParser, async (req, res, next) => {
   try {
@@ -384,7 +371,7 @@ router.post("/news/delete", urlencodedParser, async (req, res, next) => {
   }
 });
 
-router.post("/news/add", urlencodedParser, async (req, res, next) => {
+router.post("/news/add", urlencodedParser, async (_req, res, next) => {
   try {
     let [results] = await db.query(
       `INSERT INTO news VALUES (0,'','2999-01-01 00:00','')`
@@ -397,11 +384,10 @@ router.post("/news/add", urlencodedParser, async (req, res, next) => {
 });
 
 router.post("/news/edit", urlencodedParser, async (req, res, next) => {
-
-  var text = viewhelpers.EscapeQuotes(req.body.text);
+  let text = viewhelpers.EscapeQuotes(req.body.text);
   try {
     if (!viewhelpers.isDate(req.body.date)) throw new Error("Invalid date");
-    var date = req.body.date.slice(0, 19).replace("T", " ");
+    let date = req.body.date.slice(0, 19).replace("T", " ");
     let id=Number.parseInt(req.body.id);
     await db.query(
       `UPDATE news SET title = '${req.body.title}', \
@@ -430,17 +416,21 @@ router.post("/news/posterupload", urlencodedParser, async (req, res) => {
   }
 });
 
-router.get("/gallery", async (req, res) => {
-  var names = await viewhelpers.NamesOfDirFilesWOExtension(
+router.get("/gallery", async (_req, res) => {
+  let names = await viewhelpers.NamesOfDirFilesWOExtension(
     "/static/img/gallery"
   );
   res.render("admin/gallery.hbs", { names, layout: false });
 });
 
 router.post("/gallery/delete", urlencodedParser, (req) => {
-  fs.unlink(path.join(__dirname, "../", "/static/", req.body.filename));
+  //check if filename doesn't contain ../
+  if (req.body.filename.indexOf("../") !== -1) return;
+
+  let filename = req.body.filename;
+  fs.unlink(path.join(__dirname, "../", "/static/", filename));
   fs.unlink(
-    path.join(__dirname, "../", "/static/thumbnails/", req.body.filename)
+    path.join(__dirname, "../", "/static/thumbnails/", filename)
   );
 });
 
@@ -448,7 +438,7 @@ router.post("/gallery/upload", urlencodedParser, (req, res) => {
   if (!req.files) {
     return res.status(400);
   }
-  var files = FilesToArray(req.files);
+  let files = FilesToArray(req.files);
   files.forEach((fileToUpload) => {
     let tmpfile = path.join(__dirname, "..", "/tmp/", fileToUpload.name);
     fileToUpload.mv(tmpfile, function () {
@@ -497,8 +487,8 @@ router.post("/artists/delete", urlencodedParser, async (req, res, next) => {
 
 router.post("/artists/translate", urlencodedParser, async (req, res, next) => {
   let currentLang = globals.languages[req.getLocale()];
-  var id = Number.parseInt(req.body.id);
-  var sourceLang = req.getLocale();
+  let id = Number.parseInt(req.body.id);
+  let sourceLang = req.getLocale();
   try {
     let [artist] = await db.query(
       `SELECT name, instrument, country FROM artists_translate WHERE languageId=${currentLang} AND artistId=${id}`
@@ -514,10 +504,10 @@ router.post("/artists/translate", urlencodedParser, async (req, res, next) => {
         if (currentLang == langId) {
           continue;
         }
-        var destLang = globals.languages.getNameById(langId);
-        var name = await translate(artist[0].name, sourceLang, destLang);
-        var country = await translate(artist[0].country, sourceLang, destLang);
-        var instrument = await translate(
+        let destLang = globals.languages.getNameById(langId);
+        let name = await translate(artist[0].name, sourceLang, destLang);
+        let country = await translate(artist[0].country, sourceLang, destLang);
+        let instrument = await translate(
           artist[0].instrument,
           sourceLang,
           destLang
@@ -539,7 +529,7 @@ router.post("/artists/translate", urlencodedParser, async (req, res, next) => {
   }
 });
 
-router.post("/artists/add", urlencodedParser, async (req, res, next) => {
+router.post("/artists/add", urlencodedParser, async (_req, res, next) => {
   try {
     let insertQuery="START TRANSACTION; ";
     insertQuery+=`INSERT INTO artists VALUES (0,0); SELECT LAST_INSERT_ID() INTO @ID;`;
@@ -630,8 +620,8 @@ router.post(
   urlencodedParser,
   async (req, res, next) => {
     let currentLang = globals.languages[req.getLocale()];
-    var id = Number.parseInt(req.body.id);
-    var sourceLang = req.getLocale();
+    let id = Number.parseInt(req.body.id);
+    let sourceLang = req.getLocale();
     try {
       let [composer] = await db.query(
         `SELECT name, country FROM composers_translate WHERE languageId=${currentLang} AND composerId=${id}`
@@ -646,9 +636,9 @@ router.post(
           if (currentLang == langId) {
             continue;
           }
-          var destLang = globals.languages.getNameById(langId);
-          var name = await translate(composer[0].name, sourceLang, destLang);
-          var country = await translate(
+          let destLang = globals.languages.getNameById(langId);
+          let name = await translate(composer[0].name, sourceLang, destLang);
+          let country = await translate(
             composer[0].country,
             sourceLang,
             destLang
@@ -669,7 +659,7 @@ router.post(
   }
 );
 
-router.post("/composers/add", urlencodedParser, async (req, res, next) => {
+router.post("/composers/add", urlencodedParser, async (_req, res, next) => {
   try {
     let insertQuery="START TRANSACTION; ";
     insertQuery+=`INSERT INTO composers VALUES (0,0); SELECT LAST_INSERT_ID() INTO @ID;`;
@@ -693,11 +683,8 @@ router.post("/composers/add", urlencodedParser, async (req, res, next) => {
   }
 });
 
-
-
 router.post("/composers/edit", urlencodedParser, async (req, res, next) => {
   let langId = globals.languages[req.getLocale()];
-
   let isInResidence=req.body.isInResidence??0;
   try {
     let id=Number.parseInt(req.body.id);
@@ -767,8 +754,8 @@ router.post(
   urlencodedParser,
   async (req, res, next) => {
     let currentLang = globals.languages[req.getLocale()];
-    var id = Number.parseInt(req.body.id);
-    var sourceLang = req.getLocale();
+    let id = Number.parseInt(req.body.id);
+    let sourceLang = req.getLocale();
     try {
       let [musician] = await db.query(
         `SELECT name, bio FROM musicians_translate WHERE languageId=${currentLang} AND musicianId=${id}`
@@ -782,9 +769,9 @@ router.post(
         if (currentLang == langId) {
           continue;
         }
-        var destLang = globals.languages.getNameById(langId);
-        var name = await translate(musician[0].name, sourceLang, destLang);
-        var bio = await translate(musician[0].bio, sourceLang, destLang);
+        let destLang = globals.languages.getNameById(langId);
+        let name = await translate(musician[0].name, sourceLang, destLang);
+        let bio = await translate(musician[0].bio, sourceLang, destLang);
 
         bio = viewhelpers.EscapeQuotes(bio);
         updateQuery+=
@@ -800,7 +787,7 @@ router.post(
   }
 );
 
-router.post("/musicians/add", urlencodedParser, async (req, res, next) => {
+router.post("/musicians/add", urlencodedParser, async (_req, res, next) => {
   try {
     let insertQuery="START TRANSACTION; ";
     insertQuery+=`INSERT INTO musicians VALUES (0,0,0); SELECT LAST_INSERT_ID() INTO @ID;`;
@@ -822,7 +809,6 @@ router.post("/musicians/add", urlencodedParser, async (req, res, next) => {
     next(error);
   }
 });
-
 
 router.post("/musicians/edit", urlencodedParser, async (req, res, next) => {
   let langId = globals.languages[req.getLocale()];
@@ -860,15 +846,19 @@ router.post("/musicians/posterupload", urlencodedParser, async (req, res) => {
   }
 });
 
-router.get("/press", async (req, res) => {
-  var names = await viewhelpers.NamesOfDirFilesWOExtension("/static/img/press");
+router.get("/press", async (_req, res) => {
+  let names = await viewhelpers.NamesOfDirFilesWOExtension("/static/img/press");
   res.render("admin/press.hbs", { names, layout: false });
 });
 
 router.post("/press/delete", urlencodedParser, (req) => {
-  fs.unlink(path.join(__dirname, "../", "/static/", req.body.filename));
+    //check if filename doesn't contain ../
+  if (req.body.filename.indexOf("../") !== -1) return;
+  
+  let filename = req.body.filename;
+  fs.unlink(path.join(__dirname, "../", "/static/", filename));
   fs.unlink(
-    path.join(__dirname, "../", "/static/thumbnails/", req.body.filename)
+    path.join(__dirname, "../", "/static/thumbnails/", filename)
   );
 });
 
@@ -876,7 +866,7 @@ router.post("/press/upload", urlencodedParser, (req, res) => {
   if (!req.files) {
     return res.status(400);
   }
-  var files = FilesToArray(req.files);
+  let files = FilesToArray(req.files);
 
   files.forEach((fileToUpload) => {
     let tmpfile = path.join(__dirname, "..", "/tmp/", fileToUpload.name);
@@ -894,8 +884,6 @@ router.post("/press/upload", urlencodedParser, (req, res) => {
   });
   res.redirect("/admin/");
 });
-
-
 
 router.get("/archive", async (req, res, next) => {
   try {
@@ -930,9 +918,6 @@ router.get("/archive", async (req, res, next) => {
   }
 });
 
-
-
-
 router.post("/archive/delete", urlencodedParser, async (req, res, next) => {
   try {
     let id = Number.parseInt(req.body.id);
@@ -944,7 +929,7 @@ router.post("/archive/delete", urlencodedParser, async (req, res, next) => {
   }
 });
 
-router.post("/archive/add", urlencodedParser, async (req, res, next) => {
+router.post("/archive/add", urlencodedParser, async (_req, res, next) => {
   try {
     let [results] = await db.query(
       `INSERT INTO concerts VALUES (0,'', DATE_FORMAT(NOW() - INTERVAL 1 MINUTE, '%Y-%m-%d %H:%i:00'),'', '','',1)`
@@ -957,12 +942,11 @@ router.post("/archive/add", urlencodedParser, async (req, res, next) => {
 });
 
 router.post("/archive/edit", urlencodedParser, async (req, res, next) => {
-
-  var description = viewhelpers.EscapeQuotes(req.body.description);
   let hidden = typeof req.body.hidden == "undefined" ? 0 : 1;
+  let description = viewhelpers.EscapeQuotes(req.body.description);
   try {
     if (!viewhelpers.isDate(req.body.date)) throw new Error("Invalid date");
-    var date = req.body.date.slice(0, 19).replace("T", " ");
+    let date = req.body.date.slice(0, 19).replace("T", " ");
     let id = Number.parseInt(req.body.id);
     await db.query(
       `UPDATE concerts SET title = '${req.body.title}', \
@@ -992,20 +976,24 @@ router.post("/archive/posterupload", urlencodedParser, async (req, res) => {
   }
 });
 
-router.get("/disks", async (req, res) => {
-  var names = await viewhelpers.NamesOfDirFilesWOExtension("/static/img/disks");
+router.get("/disks", async (_req, res) => {
+  let names = await viewhelpers.NamesOfDirFilesWOExtension("/static/img/disks");
   res.render("admin/disks.hbs", { names, layout: false });
 });
 
 router.post("/disks/delete", urlencodedParser, (req) => {
-  fs.unlink(path.join(__dirname, "../", "/static/", req.body.filename));
+  //check if filename doesn't contain ../
+  if (req.body.filename.indexOf("../") !== -1) return;
+  
+  let filename = req.body.filename;
+  fs.unlink(path.join(__dirname, "../", "/static/", filename));
 });
 
 router.post("/disks/upload", urlencodedParser, (req, res) => {
   if (!req.files) {
     return res.status(400);
   }
-  var files = FilesToArray(req.files);
+  let files = FilesToArray(req.files);
 
   files.forEach((fileToUpload) => {
     let tmpfile = path.join(__dirname, "..", "/tmp/", fileToUpload.name);
@@ -1019,25 +1007,27 @@ router.post("/disks/upload", urlencodedParser, (req, res) => {
   res.redirect("/admin/");
 });
 
-
-router.get("/about", (req, res) => {
+router.get("/about", (_req, res) => {
   res.render("admin/about.hbs", { layout: false });
 });
 
-router.get("/conductor", (req, res) => {
+router.get("/conductor", (_req, res) => {
   res.render("admin/conductor.hbs", {layout: false });
 });
 
-router.get("/pastmusicians", (req, res) => {
+router.get("/pastmusicians", (_req, res) => {
   res.render("admin/pastmusicians.hbs", {layout: false });
 });
 
-router.get("/contacts", (req, res) => {
+router.get("/contacts", (_req, res) => {
   res.render("admin/contacts.hbs", {layout: false });
 });
 
 router.post("/updatestatichtml", async (req, res) => {
   try {
+      //check if filename doesn't contain ../
+    if (req.body.file.indexOf("../") !== -1) return;
+
     let file = req.body.file;
     let html= req.body.content;
     await fs.writeFile(path.join("static",file),html);
